@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "rm -rf build; babel src -d build --copy-files",
-    "start": "npm run-script build && node build/index.js",
+    "start": "nodemon --exec babel-node src/index.js",
     "dev": "NODE_ENV='development' npm run-script start",
     "prod": "NODE_ENV='production' npm run-script start",
     "lint": "./node_modules/.bin/eslint -c ./.eslintrc.json src/**/*.js",
@@ -14,7 +14,7 @@
     "coverage": "jest --collectCoverageFrom=./src/**/*.js --coverage src"
   },
   "engines": {
-    "node": "8.11.3",
+    "node": ">=8.11.3",
     "npm": "5.6.0"
   },
   "repository": {
@@ -33,20 +33,24 @@
   },
   "homepage": "https://github.com/vstostap/node-boilerplate#readme",
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-latest": "^6.24.1",
-    "eslint": "^5.5.0",
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/node": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "eslint": "^5.7.0",
     "eslint-plugin-es-beautifier": "^1.0.1",
-    "jest": "^23.5.0"
+    "jest": "^23.5.0",
+    "nodemon": "^1.18.4"
   },
   "dependencies": {
     "bluebird": "^3.5.2",
     "bunyan": "^1.8.12",
     "crypto": "^1.0.1",
-    "express": "^4.16.3",
+    "express": "^4.16.4",
     "http": "0.0.0",
-    "joi": "^13.6.0",
-    "mongoose": "^5.2.13"
+    "https": "^1.0.0",
+    "joi": "^14.0.0",
+    "mongoose": "^5.3.4"
   }
 }
