@@ -1,24 +1,23 @@
-const path = require('path');
-
+import path from 'path';
 // Export
 module.exports = {
   app: {
-        host: '0.0.0.0',
-        port: 3000
+    host: '0.0.0.0',
+    port: 3000
   },
   database: {
-        host: 'localhost',
-        port: 27017,
-        name: 'node-boilerplate'
+    host: 'localhost',
+    port: 27017,
+    name: 'node-boilerplate'
   },
   logs: {
-        name: 'node-boilerplate',
-        folder: path.join(__dirname, '../logs/'),
-        streams: [
-            {
-                level: 'debug',
-                stream: process.stdout // log INFO and above to stdout
-            }
-        ]
+    name   : 'node-boilerplate',
+    folder : path.join(__dirname, '../logs/'),
+    streams: [
+      {
+        level : 'debug',
+        stream: process.stdout // log INFO and above to stdout
+      }
+    ]
   }
 }
