@@ -1,20 +1,19 @@
 // Base configuration
-consrt config = require('./development');
-
+import config from './development';
 // Override configurations for Production environment
 config.app.prefix = '/api';
 
 config.logs.folder = '/var/log';
 config.logs.streams = [
-    {
-        level: 'info',
-        path: config.logs.folder + '/node-boilerplate.log'
-    },
-    {
-        level: 'error',
-        stream: process.stderr,
-        path: config.logs.folder + '/node-boilerplate.error.log'
-    }
+  {
+    level: 'info',
+    path : config.logs.folder + '/node-boilerplate.log'
+  },
+  {
+    level : 'error',
+    stream: process.stderr,
+    path  : config.logs.folder + '/node-boilerplate.error.log'
+  }
 ];
 
 // Export
