@@ -2,16 +2,15 @@
  * Imports
  */
 import config from '../config';
+import mongoose from 'mongoose';
 import log from './logging';
-
-/**
- * Import and initialize MongoDb connection
- */
-export const mongoose = require('mongoose');
 
 const connectionString = `mongodb://${config.database.host}:${config.database.port}/${config.database.name}`;
 
-mongoose.connect(connectionString, err => {
+/**
+ * Initialize MongoDb connection
+ */
+mongoose.connect(connectionString, { useNewUrlParser: true }, err => {
   if (err) return log.info(`MongoDb connection error occured ${err}`);
   log.info('Connected to MongoDb successfully');
 });
@@ -24,3 +23,5 @@ mongoose.connection.on('disconnected', () => {
   const now = new Date();
   log.info(`MongoDb disconnected at ${now}`);
 });
+
+export default mongoose;
